+++ R/get_can_land_use.R
@@ -26,10 +26,10 @@ get_can_land_use <- function(years, zones, path = ".") {
       msgInfo("year", style_bold(i), " zone", style_bold(j), " ", 
         appendLF = FALSE)
       zout <- tempfile(fileext = ".zip")
-      curl_download(paste0(luurl, tmp), destfile = zout)
+      dl_check(url = paste0(luurl, tmp), destfile = zout)
       unzip(zout, exdir = path)
       unlink(zout)
-      msgSuccess("")
+      msgSuccess("File extracted")
     }
   }